+++ CHANGELOG.md
@@ -7,7 +7,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
-The latest version contains all changes.
+<!-- The latest version contains all changes. -->
+
+### Changed
+
+- Bold table headers instead of repo names
+
+### Removed
+
+- Extra newline before first entry when printing all tables
 
 ## [1.2.0] - 2021-05-16