+++ CHANGELOG.md
@@ -7,7 +7,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
-The latest version contains all changes.
+<!-- The latest version contains all changes. -->
+
+### Changed
+
+- Config type to be embedded within the Driver
+  - Not in public library modules, but this should improve generation efficiency
 
 ## [1.2.1] - 2021-05-23