+++ CHANGELOG.md
@@ -7,7 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
-<!-- The latest version contains all changes. -->
+The latest version contains all changes.
+
+## [1.3.0] - 2021-05-25
 
 ### Changed